pyproject.toml
[tool.poetry]
name = "logdetective"
version = "3.10.0"
description = "Analyze logs with a template miner and an LLM to discover errors and suggest solutions."
authors = ["Jiri Podivin <jpodivin@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
include = [
    "logdetective/drain3.ini",
    "logdetective/server/templates/gitlab_comment.md.j2",
    "logdetective/prompts.yml",
    "logdetective.1.asciidoc",
    "logdetective/prompts/system_prompt.j2",
]
packages = [
    { include = "logdetective" }
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Internet :: Log Analysis"
]

[tool.poetry.urls]
homepage = "https://github.com/fedora-copr/logdetective"
issues = "https://github.com/fedora-copr/logdetective/issues"

[tool.poetry.dependencies]
python = "^3.11"
llama-cpp-python = ">0.2.56,!=0.2.86,<1.0.0"
drain3 = {git = "https://github.com/logpai/Drain3.git", rev = "0fb6d6e1828838fa05c8f6fa5cda878e02ea79c7"}
huggingface-hub = ">=1.0.0,<2.0.0"
# rawhide has numpy 2, F40 and F41 are still on 1.26
# we need to support both versions
numpy = ">=1.26.0"
pydantic = ">=2.8.2,<3.0.0"
pyyaml = ">=6.0.1,<7.0.0"
aiohttp = ">=3.7.4,<4.0.0"  # latest version in el8
# Protection against supply chain compromise https://docs.litellm.ai/blog/security-update-march-2026
litellm = "!=1.82.7,!=1.82.8"
python-gitlab = {version = ">=4.4.0", optional = true }
fastapi = {version = ">=0.111.1,<1.0.0", optional = true }
sqlalchemy = {version = ">=2.0.36,<3.0.0", optional = true }
alembic = {version = ">=1.13.3,<2.0.0", optional = true }
backoff = {version = "2.2.1", optional = true }
sentry-sdk = {version = ">=2.17.0,<3.0.0", optional = true, extras = ["fastapi"]}
koji = {version = ">=1.35.0,<2.0.0", optional = true}
responses = {version = ">=0.25.7,<0.26.0", optional = true}
aioresponses = {version = ">=0.7.8,<0.8.0", optional = true}
pytest-asyncio = {version = ">=1.1.0,<2.0.0", optional = true}
pytest-mock = {version = ">=3.14.1,<4.0.0", optional = true}
flexmock = {version = ">=0.12.2,<0.13.0", optional = true}
pytest = {version = ">=8.4.1,<9.0.0", optional = true}
pytest-cov = {version = "^7.0.0", extras = ["testing"], optional = true}
asyncpg = {version = ">=0.30.0,<1.0.0", optional = true}
asciidoc = {version = "^10.2.1", optional = true, extras = ["testing"]}
httpx = {version = ">=0.28.1,<1.0.0", optional = true}
beeai-framework = ">=0.1.78"
starlette = {version = "<1.0.0", optional = true }
uvicorn = {version = "^0.44.0", optional = true }
gunicorn = {version = "^25.3.0",  optional = true }

[tool.poetry.extras]
server = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "backoff",
    "sentry-sdk",
    "koji",
    "asyncpg",
    "python-gitlab",
    "httpx",
    "starlette",
    "uvicorn",
    "gunicorn",
]
server-testing = [
    "fastapi",
    "sqlalchemy",
    "alembic",
    "backoff",
    "sentry-sdk",
    "koji",
    "pytest-mock",
    "responses",
    "asyncpg",
    "python-gitlab",
    "httpx",
    "starlette",
    "uvicorn",
    "gunicorn",
]
testing = [
    "pytest",
    "pytest-asyncio",
    "aioresponses",
    "flexmock",
    "pytest-cov",
    "asciidoc",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
logdetective = 'logdetective.logdetective:main'

[tool.pylint]
disable = [
    "inconsistent-return-statements",
    "missing-module-docstring",
    "too-few-public-methods",
    "unspecified-encoding",
]
extension-pkg-allow-list = ["socket"]

[tool.pylint.typecheck]
# TODO: Pylint fails in CI on loading `AddressFamily` enum from socket,
# other approaches have failed to resolve the issue.
# Remove when the https://github.com/pylint-dev/pylint/issues/9850 is fixed.
# Tell pylint to ignore 'no-member' errors for these specific attributes
# because it cannot reliably find them in the C-extension statically.
generated-members = ["socket.AddressFamily", "AddressFamily"]


[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.djlint]
profile="jinja"
extension=".j2"
